Add format email function
